python_influx/requirements.txt
PyMySQL
influxdb
websockets
numpy
pika==1.0.0
psycopg2-binary
requests_futures
